Safari-609.1.11.1/LayoutTests/fast/dom/HTMLFontElement/face-attribute-expected.txt

HTMLFontElement face attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fontFaceAttributeEffect("") is null
PASS fontFaceAttributeEffect(" ") is null
PASS fontFaceAttributeEffect(",") is null
PASS fontFaceAttributeEffect(" ,") is null
PASS fontFaceAttributeEffect(" , ") is null
PASS fontFaceAttributeEffect(",,") is null
PASS fontFaceAttributeEffect("serif") is "serif"
PASS fontFaceAttributeEffect("serif ") is "serif"
PASS fontFaceAttributeEffect(",serif,") is null
PASS fontFaceAttributeEffect(" serif ") is "serif"
PASS fontFaceAttributeEffect(" serif") is "serif"
PASS fontFaceAttributeEffect("serif ,") is null
PASS fontFaceAttributeEffect("serif , ") is null
PASS fontFaceAttributeEffect("serif,,") is null
PASS fontFaceAttributeEffect("Serif") is "serif"
PASS fontFaceAttributeEffect(" Serif") is "serif"
PASS fontFaceAttributeEffect(" Serif ") is "serif"
PASS fontFaceAttributeEffect(",Serif ") is null
PASS fontFaceAttributeEffect("Inherited") is "Inherited"
PASS fontFaceAttributeEffect(",foo + foo,serif") is null
PASS fontFaceAttributeEffect("serif,foo + foo,") is null
PASS successfullyParsed is true

TEST COMPLETE
